Added correct and working code for BFS on a composed graph (TS x DFA). Tested it for various formulas and the implementation seems to be working
NOTE: During some runs, I did observe the cube extraction process does not output a complete string, i.e. a valid assignment for a given bdd is not same as the # of variables relevant to that boolean formula. This is could be memory related issues. Need some robust way of detecting and eliminating such incidents.