WIP: Added a work in progress preference based symbolic plannign algorithm based on peter's ICRA23 algorithm .
This brnach exisits for book-keeping purposes only. I may come back in the future and add this feature.

src/algorithms/weighted_search/preference_BDDAstar.py - Added skeleton code to compute the PCV value for each state. The code still needs to validated and verified for its behavior. Currently, I am enumerating all the states in my ADD; this has proven to be slow in the past. I need to find a quiker way of doing it.