Added code for human to not intervene and move a box at support loc with another box on top.
Valid Human Moves for Arch Problem: Human can move any object placed on the top. A human can only move an object placed at the support location if there is no object placed on top.

NOTE: For the arch construction formula, 'F(p00 & p12 & p21 & free) & G(~(p00 & p21) -> ~(p12))', with 0 human interventions, we are able to synthesize winning strategies for the robot. When we allow the human to Intervene, there does not exists a strategy. Need to find and fix this.