WIP: Adding code for no-preds. TR Edge construction fixed.
Important Note: As we are using the Compositional approach we can not have ambiguity in state encoding, i.e., each boolean variable corresponding to a state has to be either true or false. Currently, our state is defined as (robot-conf).(world-conf) --- (a_s & a_e) --->  (robot-conf)'(world-conf)'. Each boolean variable in (robot-conf) and (robot-conf)' needs to be fully defined and also in the (world-conf) and (world-conf)'.

For this reason, when creating the different sets of Boolean vars for Robot and World Conf, I had to explicitly create (not ready), (not holding), (not to-obj) and (not to-loc) for predicates of type World Conf. Similarly, I created (not b0), (not b1), ... for all boxes to indicate that box `b_i` is currently being manipulated by the arm. This, is done to unambiguously describe each states.

franka_strategy_main.py: Added code to create `not-predicates` of type world conf and robot conf. Refer above notes for details.

symbolic_game_abstraction.py: Added function get_missing_preds() for loop through a state encoding, find the type of predicate that is missing and manually add `not-predicate` for that predicate.