Adding Updated abstraction construction for Bounded and Unbounded Scenarios.
Implementation Notes:

Bounded Abstraction: The abstraction is similar to the one used by Keliang. From each state in the TS, you have a robot edge and a human edge. You either evolve accoridng to human edge or robot edge. The winning strategy roll-out functionality is thus different. During roll out, I toss a coin and choose for either the human to intervene or not intervene (evolve accoridng to Robot's action).

Unbounded Abstraction: The above abstraction interpretation does not work due to the unbounded nature of human interventions. For e.g., at a given state say the robot wants to `release` the box in End Effector but the human keeps moving some other block. Using the above interpretation, the graph will always evolve as per the human move and thus the robot won't be able to `release` the box.

['(on b0 ee)', '(on b1 l6)', '(holding l0)'] -------(release b0 l0) (human-move b1 l6 l7)------> ['(on b0 ee)', '(on b1 l7)', '(holding l0)']
['(on b0 ee)', '(on b1 l7)', '(holding l0)'] -------(release b0 l0) (human-move b1 l7 l6)------> ['(on b0 ee)', '(on b1 l6)', '(holding l0)']

To resolve this, in this abstraction, the graph evolves as per the combined robot and human action, i.e., (TS-state) ---- (release) & (human move) ---> (TS-state)'. Here (TS-state)' reflects robot's intention to perform an action (release) and the actual evolution as per the human's action  (move some other block). This interpretation leads to following evolution

['(on b0 ee)', '(on b1 l6)', '(holding l0)'] -------(release b0 l0) (human-move b1 l6 l7)------> ['(on b0 l0)', '(on b1 l7)', '(ready l0)']

The roll-out of the winning strategy is thus different for Unbounded Reachability game.

Note: The strategy synthesis is exactly the same for Bounded and Unbounded Reachability game (except for additional Human Intervention Variables).