BUG FIX: Kindof fixed Cooperative strategy rollout. Updated function call in the main module
Details about Cooperative strategy rollout: The cooperative strategy rollout sometime fails to compute a valid human action.

Reason: Primarily, this is due to the way we build `non-intervening` human edges (in our Abstraction) which are a compliment of the valid human edge(s) from a given state. During Min-Min synthesis a non-intervening and intervening edge could both be equally optimal. Thus, during the computation of possible human action for a given state and robot action pair, we get human action(s) that are not optimal. This is something to do with Human_act_cube synthesis and should be resolved in future iterations.

For sanity checking I also print optimal values associated with states during the rollout. We can manually check the evolution under (s, a_s, a_e).

To reiterate, the strategy synthesis and the Min. energy required in the Min-Min Game are correct!