Feature: Added code to store TR accordingly org action names instead of modified actions.
Details: During our strategy synthesis process, we use the index that a TR is stored in, to determine the weight assocated with the corresponding action. To reason over each paramterized action (much needed during qunatitative games), I modified the implementation to create Symbolic Transition Functions and store in a list which is equal to the length of number of original actions.